Add : Mission 1 - Id, Password input 상태 관리 함수 구현
