--- src/pages/mihye/Login/login.scss
@@ -1,95 +1,0 @@
-.login {
-  width: 100%;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: #f5f5f5;
-}
-
-.wrap {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  border-style: solid;
-  border-color: #ededed;
-  border-width: 2px;
-  background-color: white;
-}
-
-.logo {
-  margin-top: 50px;
-  margin-bottom: 30px;
-  text-align: center;
-  font-size: 50px;
-  font-family: "Lobster", cursive;
-}
-
-.login_box {
-  margin: 30px;
-}
-
-.id_box {
-  margin-bottom: 7px;
-}
-
-input::placeholder {
-  color: grey;
-  font-size: 12px;
-  font-weight: 500;
-  padding-left: 7px;
-}
-
-.input_id,
-.input_password {
-  width: 300px;
-  height: 36px;
-  background-color: #f5f5f5;
-  border: solid;
-  border-width: 2px;
-  border-color: #ededed;
-  border-radius: 3px;
-}
-
-.password_box {
-  color: #f5f5f5;
-  margin-bottom: 15px;
-}
-
-.login_button {
-  width: 308px;
-  height: 36px;
-  border: none;
-  border-radius: 5px;
-  background-color: #4697f7;
-  color: white;
-  font-size: 14px;
-  font-weight: 600;
-}
-
-.login_button:hover:enabled {
-  background-color: #4697f7;
-}
-
-.login_button:disabled {
-  opacity: 0.4;
-  cursor: not-allowed;
-}
-
-.bottom_text {
-  margin: 65px;
-  font-size: 14px;
-  color: #00489d;
-
-  font-family: "Noto Sans KR", sans-serif;
-}
-
-@media (max-width: 2500px) {
-  .wrap {
-    width: 400px;
-    height: 480px;
-    padding: 10px;
-    font-weight: 600;
-  }
-}
