+++ src/pages/mihye/Main/Main.scss
@@ -1,0 +1,192 @@
+* {
+  margin: 0;
+  padding: 0;
+}
+
+ul li {
+  list-style: none;
+}
+.cotainer {
+  margin: 30px 60px;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+
+  #navbar {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background-color: white;
+
+    .navbar_home {
+      display: flex;
+      align-items: center;
+
+      .navber_home_title_image {
+        object-fit: none;
+        width: 150px;
+        height: auto;
+      }
+
+      .navber_home_title_image {
+        img {
+          width: 150px;
+          height: auto;
+        }
+      }
+
+      .navbar_home_vertical_line {
+        width: 5px;
+        height: 40px;
+        border-left: solid;
+        border-color: black;
+        border-width: 2px;
+        margin-left: 9px;
+        margin-bottom: 4.5px;
+      }
+    }
+
+    .search_box {
+      width: 217px;
+      height: 25px;
+      border-radius: 3px;
+      border: 1px solid grey;
+      background-color: #f5f5f5;
+    }
+
+    input[name="search_box"]::placeholder {
+      color: grey;
+      opacity: 1;
+      text-align: center;
+    }
+
+    input[name="search_box"]::-webkit-input-placeholder {
+      background-image: url("../../../assets/logo_image/search.png");
+      background-size: contain;
+      background-position: 1px center;
+      background-repeat: no-repeat;
+      text-align: center;
+      text-indent: 0;
+    }
+
+    ul,
+    li {
+      display: inline-block;
+    }
+
+    .navbar_icon_box {
+      margin: 0 5px;
+    }
+  }
+}
+
+.main {
+  display: flex;
+  justify-content: center;
+  padding-top: 40px;
+
+  .feeds {
+    border: solid 3px #ededed;
+    border-radius: 4px;
+
+    article {
+      background-color: white;
+      margin: 20px 5px;
+      background-color: white;
+
+      .user_profile_box {
+        display: flex;
+        align-items: center;
+        margin: 10px;
+        font-size: 20px;
+
+        .user_profile_img {
+          width: 50px;
+        }
+      }
+
+      .feed_image img {
+        width: 100%;
+      }
+
+      .feed_icon_box {
+        margin-bottom: 2px;
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+
+        .feed_icon_left_box img {
+          margin: 4px 10px;
+        }
+
+        .feed_icon_right_box img {
+          margin: 4px 10px;
+        }
+      }
+
+      .who_liked {
+        font-size: 20px;
+        margin: 3px 10px;
+      }
+
+      .comment_box {
+        background-color: white;
+        font-size: 20px;
+        margin: 14px 10px;
+      }
+    }
+  }
+  .main_right_box {
+    margin-left: 30px;
+    width: 260px;
+
+    .my_account_profile_wrap {
+      display: flex;
+      align-items: center;
+
+      .my_account_profile_image {
+        width: 50px;
+      }
+    }
+
+    .story_list {
+      border: solid 3px #ededed;
+      margin: 30px 0;
+    }
+
+    .commend_list {
+      border: solid 3px #ededed;
+    }
+  }
+}
+
+.comment_input {
+  border: solid 2px #ededed;
+
+  width: 100%;
+  height: 50px;
+}
+@media (max-width: 768px) {
+  .cotainer {
+    margin: 30px 30px;
+  }
+}
+
+@media (max-width: 704px) {
+  .navber_home_title_image {
+    display: none;
+  }
+}
+
+@media (max-width: 555px) {
+  .navber_home_logo_image,
+  .navbar_home_vertical_line {
+    display: none;
+  }
+}
+
+@media (max-width: 900px) {
+  .main_right_box {
+    display: none;
+  }
+}
